Move tag-admin sources & tests to /tag-admin/src/
Import paths are modified to use relative, fully qualified file names.
