+++ .codecov.yml
@@ -13,6 +13,6 @@ flags:
 
   non-frontend-js:
     paths:
-      - tests/**.js
+      - tag-admin/src/**.js
       - pontoon/static/js/
     carryforward: true
