+++ .github/workflows/js-lint.yml
@@ -54,4 +54,4 @@ jobs:
       - name: Install dependencies
         run: npm install
       - name: prettier
-        run: ./node_modules/.bin/prettier --check '**/frontend/**/*.{js,ts,tsx,css}' '**/pontoon/**/*.{js,css}' '**/tests/**/*.{js,css}'
+        run: ./node_modules/.bin/prettier --check '**/frontend/**/*.{js,ts,tsx,css}' '**/pontoon/**/*.{js,css}' '**/tag-admin/src/**/*.{js,css}'
