+++ tag-admin/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import TagResourcesButton from './button';
+import TagResourcesButton from './button.js';
 
 document.addEventListener('DOMContentLoaded', () => {
     document.querySelectorAll('.js-tag-resources').forEach((node) => {
