+++ tag-admin/src/utils/strip.js
