+++ tag-admin/src/widgets/checkbox.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { mount, shallow } from 'enzyme';
 
-import Checkbox from 'widgets/checkbox';
+import Checkbox from './checkbox.js';
 
 test('Checkbox render', () => {
     let checkbox = shallow(<Checkbox />);
